refactor: 数据库中的tinyint字段映射到java对象类型由Byte修改为Integer refactor: user服务领域对象中使用到的状态字段修改为枚举类型，dto中使用枚举msg，do中使用枚举value

+++ wall-user-srv/src/main/java/com/oyoungy/ddd/infra/dao/UserDAO.java
@@ -17,6 +17,6 @@ public interface UserDAO extends JpaRepository<UserDO, Long> {
     @Modifying
     @CanIgnoreReturnValue
     @Query(value = "update UserDO set online = :online, lastLogin = :lastLogin where id = :userId")
-    int update4Login(@Param("userId") long userId, @Param("online") byte online, @Param("lastLogin") Date lastLogin);
+    int update4Login(@Param("userId") long userId, @Param("online") int online, @Param("lastLogin") Date lastLogin);
 
 }
